+++ shotmanager/operators/general.py
@@ -41,7 +41,7 @@ _logger = sm_logging.getLogger(__name__)
 
 class UAS_ShotManager_OT_ShotsPlayMode(Operator):
     bl_idname = "uas_shot_manager.shots_play_mode"
-    bl_label = "Shot Manager - Toggle Shots Play Mode"
+    bl_label = "Ubisoft Shot Mng - Toggle Shots Play Mode"
     bl_description = "Enable / disable the Shots Play Mode"
     bl_options = {"INTERNAL"}
 
@@ -55,7 +55,7 @@ class UAS_ShotManager_OT_ShotsPlayMode(Operator):
 
 class UAS_ShotManager_OT_DisplayOverlayTools(Operator):
     bl_idname = "uas_shot_manager.display_overlay_tools"
-    bl_label = "Shot Manager - Toggle Overlay Tools Display"
+    bl_label = "Ubisoft Shot Mng - Toggle Overlay Tools Display"
     bl_description = "Show or hide the Sequence Timeline, Interactive Shots Stack and some other tools"
     bl_options = {"INTERNAL"}
 
@@ -83,7 +83,7 @@ class UAS_ShotManager_OT_DisplayDisabledShotsInOverlays(Operator):
     # bl_description = "Display Disabled Shots in Overlay Tools"
     bl_options = {"INTERNAL"}
 
-    def invoke(self, context, event):
+    def execute(self, context):
         props = context.scene.UAS_shot_manager_props
 
         val = not props.interactShotsStack_displayDisabledShots